+++ CHANGELOG.md
@@ -12,9 +12,11 @@
 - `Innmind\Filesystem\Exception\FailedToLoadFile`
 - `Innmind\Filesystem\Exception\FailedToWriteFile`
 
-### Fixed
+## 7.6.0 - 2025-03-21
 
-- Support for PHP `8.4`
+### Added
+
+- Support for `innmind/black-box` `6`
 
 ## 7.5.1 - 2024-03-27